refactor data_reader module to kepp the structure more consistent
